Update the examples folder readme
