+++ examples/quickstart/requirements.txt
@@ -1,0 +1,3 @@
+bentoml
+torch
+transformers
