+++ examples/sk-learn/bentofile.yaml
@@ -1,0 +1,13 @@
+service: "service.py:IrisClassifier"
+labels:
+  owner: bentoml-team
+  project: gallery
+include:
+  - "*.py"
+python:
+  packages:
+    - scikit-learn
+    - pandas
+envs:
+- name: HF_TOKEN
+  value: xxx
